+++ lib/src/bot_runner.dart
@@ -6,8 +6,6 @@ class BotRunner {
   final ChatterboxStore dialogDao;
   final List<Flow> flows;
 
-  // ... (other class members and methods)
-
   BotRunner({
     required this.botToken,
     required this.dialogDao,