+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.1
 
 - Add verification for InlineButtons
+- Process callback for InlineButtons
 
 ## 1.0.0